requirements.txt
streamlit
easyocr
Pillow
chromadb
sentence-transformers
google-generativeai